internal: migrate the GPUI shell onto HeroGPUI (#17)
* internal: migrate the GPUI shell onto HeroGPUI

Shared controls now consume the published HeroGPUI facade and theme recipes so the native shell tracks HeroUI 3.2.4 instead of a second hand-rolled widget set.


* internal: close remaining GPUI/Electron chrome drift

Put shared overlay, chip, and preview look on theme recipes and reusable
helpers so windows compose chrome instead of painting it at every call site.


* internal: fold remaining GPUI views onto shared icon buttons

Settings, wallpaper, editor zoom, and video-editor panels now compose
icon_button helpers so compact-icon chrome is no longer painted per view.


* internal: satisfy Rust 1.98 clippy on the native daemons


* internal: unblock Windows GPUI and Linux daemon compiles


* internal: satisfy platform-specific GPUI clippy and test cfgs


* internal: treat Windows AsyncAppContext::update as infallible


* internal: allow HeroGPUI git and bzip2 in cargo-deny


* internal: account for the Windows drag-strip inset in title tests


* internal: keep the Windows drag inset helper live on every target


---------

+++ src/main/app-gpui/src/system/work_area.rs
@@ -13,11 +13,12 @@
 //! the taskbar (and any other appbar).
 
 use gpui::{px, Bounds, Pixels};
+use herogpui::gpui;
 
 pub fn display_bounds(display: &dyn gpui::PlatformDisplay) -> Bounds<Pixels> {
     #[cfg(target_os = "macos")]
     unsafe {
-        let bounds = CGDisplayBounds(u32::from(display.id()));
+        let bounds = CGDisplayBounds(u64::from(display.id()) as u32);
         Bounds {
             origin: gpui::point(px(bounds.origin.x as f32), px(bounds.origin.y as f32)),
             size: gpui::size(px(bounds.size.width as f32), px(bounds.size.height as f32)),
@@ -29,7 +30,7 @@ pub fn display_bounds(display: &dyn gpui::PlatformDisplay) -> Bounds<Pixels> {
 
 pub fn capture_scale_factor(display: &dyn gpui::PlatformDisplay, cx: &mut gpui::App) -> f32 {
     #[cfg(target_os = "macos")]
-    if let Some(scale) = macos_display_scale_factor(u32::from(display.id())) {
+    if let Some(scale) = macos_display_scale_factor(u64::from(display.id()) as u32) {
         return scale;
     }
     #[cfg(windows)]
@@ -353,6 +354,7 @@ fn monitor_rects(_display: Bounds<Pixels>) -> Option<(Rect, Rect)> {
 mod tests {
     use super::*;
     use gpui::{point, size};
+    use herogpui::gpui;
 
     fn bounds(x: f32, y: f32, w: f32, h: f32) -> Bounds<Pixels> {
         Bounds {
